优化功能

+++ src/Gotrays.Rcl/Pages/Index.razor.cs
@@ -12,7 +12,6 @@ public partial class Index
 
     public ChannelDto UpdateChannelDto { get; set; }
 
-    public GiteeReleaseDto GiteeReleaseDto { get; set; }
     
     private StringNumber _selectedItem;
 
@@ -31,6 +30,10 @@ public partial class Index
             {
                 SelectChannel = _channels.FirstOrDefault(x => x.Id.ToString() == value.ToString());
             }
+            else
+            {
+                _selectedItem = _channels.FirstOrDefault()?.Id.ToString();
+            }
         }
     }
 
@@ -47,8 +50,6 @@ public partial class Index
     protected override async Task OnInitializedAsync()
     {
         await LoadChannel();
-
-        _ = UpdateAsync();
 
         if (_channels.Count == 0)
         {
@@ -70,19 +71,6 @@ public partial class Index
         SelectedItem = _channels.FirstOrDefault()?.Id.ToString();
     }
 
-    private async Task UpdateAsync()
-    {
-        GiteeReleaseDto = await AppService.GetApp();
-        
-        
-        Assembly assembly = typeof(About).Assembly;
-        if (GiteeReleaseDto.tag_name != assembly.GetName().Version.ToString())
-        {
-            updateVersion = true;
-            await InvokeAsync(StateHasChanged);
-        }
-    }
-
     private void CreateAsync()
     {
         addChannel = true;
@@ -97,6 +85,10 @@ public partial class Index
 
     private async Task DeleteChannel(Guid id)
     {
+        if (_channels.Count == 1)
+        {
+            return;
+        }
         await ChannelService.DeleteAsync(id);
         await LoadChannel();
     }
